factor io config into common code
