+++ src/common/io-config/Cargo.toml
@@ -1,0 +1,14 @@
+[dependencies]
+common-error = {path = "../error", default-features = false}
+pyo3 = {workspace = true, optional = true}
+serde = {workspace = true}
+serde_json = {workspace = true}
+
+[features]
+default = ["python"]
+python = ["dep:pyo3", "common-error/python"]
+
+[package]
+edition = {workspace = true}
+name = "common-io-config"
+version = {workspace = true}
