+++ src/common/io-config/src/lib.rs
@@ -1,0 +1,6 @@
+#[cfg(feature = "python")]
+pub mod python;
+
+pub mod config;
+
+pub use config::{AzureConfig, GCSConfig, IOConfig, S3Config};
