+++ src/daft-dsl/Cargo.toml
@@ -1,6 +1,7 @@
 [dependencies]
 bincode = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
+common-io-config = {path = "../common/io-config", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 daft-io = {path = "../daft-io", default-features = false}
 pyo3 = {workspace = true, optional = true}
@@ -10,7 +11,7 @@ serde_json = {workspace = true}
 
 [features]
 default = ["python"]
-python = ["dep:pyo3", "dep:pyo3-log", "common-error/python", "daft-core/python", "daft-io/python"]
+python = ["dep:pyo3", "dep:pyo3-log", "common-error/python", "daft-core/python", "daft-io/python", "common-io-config/python"]
 
 [package]
 edition = {workspace = true}
