+++ src/daft-dsl/src/python.rs
@@ -9,7 +9,7 @@ use daft_core::{
     python::{datatype::PyDataType, field::PyField, schema::PySchema},
 };
 
-use daft_io::python::IOConfig;
+use common_io_config::python::IOConfig as PyIOConfig;
 use pyo3::{
     exceptions::PyValueError,
     prelude::*,
@@ -358,7 +358,7 @@ impl PyExpr {
         max_connections: i64,
         raise_error_on_failure: bool,
         multi_thread: bool,
-        config: Option<IOConfig>,
+        config: Option<PyIOConfig>,
     ) -> PyResult<Self> {
         if max_connections <= 0 {
             return Err(PyValueError::new_err(format!(
