+++ src/daft-io/Cargo.toml
@@ -11,6 +11,7 @@ azure_storage = {version = "0.13.0", features = ["enable_reqwest_rustls"], defau
 azure_storage_blobs = {version = "0.13.1", features = ["enable_reqwest_rustls"], default-features = false}
 bytes = {workspace = true}
 common-error = {path = "../common/error", default-features = false}
+common-io-config = {path = "../common/io-config", default-features = false}
 daft-core = {path = "../daft-core", default-features = false}
 futures = {workspace = true}
 google-cloud-storage = {version = "0.13.0", default-features = false, features = ["rustls-tls", "auth"]}
@@ -35,7 +36,7 @@ tempfile = "3.7.1"
 
 [features]
 default = ["python"]
-python = ["dep:pyo3", "dep:pyo3-log", "common-error/python", "daft-core/python"]
+python = ["dep:pyo3", "dep:pyo3-log", "common-error/python", "daft-core/python", "common-io-config/python"]
 
 [package]
 edition = {workspace = true}
